fix: trying with full e2e test cases
